requirements.txt
Flask==1.0.2
gunicorn==19.9.0
sklearn==0.19.1
mysql-connector==2.2.9
